fix all register controller bugs

+++ src/utils/cloudinary.js
@@ -15,11 +15,14 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type: "auto"
         })
         // File uploaded successfully
-        console.log("File uploaded successfully", response.url);
+        // console.log("File uploaded successfully", response.url);
+        fs.unlinkSync(localFilePath); // remove the localluy saved temporary file after successful upload
         return response.url;
     } 
     catch (error) {
-        fs.unlinkSync(localFilePath); // remove the localluy saved temporary file as the upload failed
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath); // remove the localluy saved temporary file as the upload failed
+        }
         console.error("Error uploading file to Cloudinary", error);
         return null;
     }
